FIR/UAST: commonize UBlockExpression
