--- uast/uast-kotlin-base/src/org/jetbrains/uast/kotlin/expressions/KotlinAbstractUBlockExpression.kt
@@ -1,19 +1,0 @@
-/*
- * Copyright 2010-2021 JetBrains s.r.o. and Kotlin Programming Language contributors.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
- */
-
-package org.jetbrains.uast.kotlin
-
-import org.jetbrains.kotlin.psi.KtBlockExpression
-import org.jetbrains.uast.UBlockExpression
-import org.jetbrains.uast.UElement
-
-abstract class KotlinAbstractUBlockExpression(
-    override val sourcePsi: KtBlockExpression,
-    givenParent: UElement?
-) : KotlinAbstractUExpression(givenParent), UBlockExpression, KotlinUElementWithType {
-    override val expressions by lz {
-        sourcePsi.statements.map { baseResolveProviderService.baseKotlinConverter.convertOrEmpty(it, this) }
-    }
-}
