--- uast/uast-kotlin-fir/src/org/jetbrains/uast/kotlin/expressions/FirKotlinUBlockExpression.kt
@@ -1,17 +1,0 @@
-/*
- * Copyright 2010-2021 JetBrains s.r.o. and Kotlin Programming Language contributors.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
- */
-
-package org.jetbrains.uast.kotlin.expressions
-
-import org.jetbrains.kotlin.psi.KtBlockExpression
-import org.jetbrains.uast.UElement
-import org.jetbrains.uast.kotlin.KotlinAbstractUBlockExpression
-
-class FirKotlinUBlockExpression(
-    override val sourcePsi: KtBlockExpression,
-    givenParent: UElement?
-) : KotlinAbstractUBlockExpression(sourcePsi, givenParent) {
-    // TODO: either override convertParent or commonize it
-}
